+++ sdk/search/azure-search-documents/azure/search/documents/indexes/models/_models.py
@@ -26,7 +26,7 @@ from .._generated.models import (
     AzureActiveDirectoryApplicationCredentials,
     CognitiveServicesAccount,
     SearchIndexerKnowledgeStore,
-    SearchIndexerIndexProjections,
+    SearchIndexerIndexProjection,
 )
 
 DELIMITER = "|"
@@ -50,7 +50,7 @@ class SearchIndexerSkillset:
      enriched data.
     :vartype knowledge_store: ~azure.search.documents.indexes.models.SearchIndexerKnowledgeStore
     :ivar index_projections: Definition of additional projections to secondary search index(es).
-    :vartype index_projections: ~azure.search.documents.indexes.models.SearchIndexerIndexProjections
+    :vartype index_projection: ~azure.search.documents.indexes.models.SearchIndexerIndexProjection
     :ivar e_tag: The ETag of the skillset.
     :vartype e_tag: str
     :ivar encryption_key: A description of an encryption key that you create in Azure Key Vault.
@@ -72,7 +72,7 @@ class SearchIndexerSkillset:
         description: Optional[str] = None,
         cognitive_services_account: Optional["CognitiveServicesAccount"] = None,
         knowledge_store: Optional["SearchIndexerKnowledgeStore"] = None,
-        index_projections: Optional["SearchIndexerIndexProjections"] = None,
+        index_projection: Optional["SearchIndexerIndexProjection"] = None,
         e_tag: Optional[str] = None,
         encryption_key: Optional["SearchResourceEncryptionKey"] = None,
         **kwargs: Any
@@ -83,7 +83,7 @@ class SearchIndexerSkillset:
         self.skills = skills
         self.cognitive_services_account = cognitive_services_account
         self.knowledge_store = knowledge_store
-        self.index_projections = index_projections
+        self.index_projection = index_projection
         self.e_tag = e_tag
         self.encryption_key = encryption_key
 
@@ -101,7 +101,7 @@ class SearchIndexerSkillset:
             skills=generated_skills,
             cognitive_services_account=getattr(self, "cognitive_services_account", None),
             knowledge_store=getattr(self, "knowledge_store", None),
-            index_projections=getattr(self, "index_projections", None),
+            index_projection=getattr(self, "index_projection", None),
             e_tag=getattr(self, "e_tag", None),
             encryption_key=getattr(self, "encryption_key", None),
         )